mv relevant from global to WalletManager
